# packages/mobile/packages/server-auth/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WeChat Clone — Auth Service API",
    "version": "0.1.0",
    "description": "用户认证服务 — 注册、登录、Token 管理、用户信息管理"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "本地开发环境"
    }
  ],
  "tags": [
    { "name": "auth", "description": "认证相关接口" },
    { "name": "users", "description": "用户信息相关接口" }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "用户注册",
        "description": "注册新账户。密码使用 bcrypt 哈希存储，最少 8 位。开发阶段可选跳过验证码。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 0 },
                    "user": { "$ref": "#/components/schemas/UserProfile" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误或用户名已存在",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "用户登录",
        "description": "验证用户名密码，签发 JWT (access_token 15分钟 + refresh_token 7天)，写入 Redis Session。限流: 5次/分钟/IP。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "用户名或密码错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "用户登出",
        "description": "清除当前 Session，更新用户状态为离线。",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "登出成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 0 },
                    "message": { "type": "string", "example": "已登出" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "刷新 Token",
        "description": "使用 refresh_token 换取新的 access_token 和 refresh_token。旧的 refresh_token 将被作废。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "description": "登录时获取的刷新令牌"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "刷新成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": {
            "description": "缺少 refreshToken 参数",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Token 无效或已过期",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "tags": ["users"],
        "summary": "获取当前用户信息",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "当前用户完整信息",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 0 },
                    "user": { "$ref": "#/components/schemas/UserProfile" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "更新个人资料",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProfileRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 0 },
                    "user": { "$ref": "#/components/schemas/UserProfile" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/users/me/avatar": {
      "put": {
        "tags": ["users"],
        "summary": "上传头像",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "nullable": true,
                    "description": "头像 URL，传 null 清除头像"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 0 },
                    "user": { "$ref": "#/components/schemas/UserProfile" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me/password": {
      "put": {
        "tags": ["users"],
        "summary": "修改密码",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码修改成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 0 },
                    "message": { "type": "string", "example": "密码修改成功" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "新密码强度不足",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "原密码错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["users"],
        "summary": "获取用户公开信息",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "用户 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "用户公开信息",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 0 },
                    "user": { "$ref": "#/components/schemas/UserPublicProfile" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "用户不存在",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/users/search": {
      "post": {
        "tags": ["users"],
        "summary": "搜索用户",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "搜索关键词 (按用户名或昵称模糊匹配)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "搜索结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 0 },
                    "items": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/UserPublicProfile" }
                    },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": ["auth"],
        "summary": "健康检查",
        "responses": {
          "200": {
            "description": "服务正常",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "service": { "type": "string", "example": "auth" },
                    "timestamp": { "type": "integer", "example": 1714000000000 }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "在 Authorization 头中传入 access_token: Bearer <token>"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "password", "nickname"],
        "properties": {
          "username": { "type": "string", "description": "用户名 (字母开头, 3-20位)", "minLength": 3, "maxLength": 20, "pattern": "^[a-zA-Z][a-zA-Z0-9_]{2,19}$" },
          "password": { "type": "string", "description": "密码 (最少8位)", "minLength": 8, "maxLength": 128 },
          "nickname": { "type": "string", "description": "昵称 (1-50位)", "minLength": 1, "maxLength": 50 },
          "phone": { "type": "string", "description": "手机号 (可选)", "pattern": "^1[3-9]\\d{9}$" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "description": "用户名" },
          "password": { "type": "string", "description": "密码" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 0 },
          "token": { "type": "string", "description": "访问令牌 (15分钟有效)" },
          "refreshToken": { "type": "string", "description": "刷新令牌 (7天有效)" },
          "expiresIn": { "type": "integer", "description": "过期时间 (秒)", "example": 900 },
          "user": { "$ref": "#/components/schemas/UserProfile" }
        }
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "nickname": { "type": "string" },
          "avatar": { "type": "string", "nullable": true },
          "phone": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["online", "offline", "away", "busy"] },
          "lastSeenAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "UserPublicProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "nickname": { "type": "string" },
          "avatar": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["online", "offline", "away", "busy"] }
        }
      },
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "nickname": { "type": "string", "description": "新昵称" },
          "avatar": { "type": "string", "description": "头像 URL" },
          "phone": { "type": "string", "description": "手机号 (传 null 清除)" }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "required": ["oldPassword", "newPassword"],
        "properties": {
          "oldPassword": { "type": "string", "description": "原密码" },
          "newPassword": { "type": "string", "description": "新密码 (最少8位)", "minLength": 8 }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "description": "错误码 (0=成功, 1xxx=通用, 2xxx=认证)" },
          "message": { "type": "string", "description": "错误描述" },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string" },
                "message": { "type": "string" }
              }
            },
            "description": "字段级校验错误 (可选)"
          }
        }
      }
    }
  }
}